STCOM-801: Remove dublication code
