+++ lib/ControlledVocab/tests/ControlledVocab-test.js
@@ -6,11 +6,17 @@ import {
   it,
 } from '@bigtest/mocha';
 
+import React from 'react';
 import ControlledVocabInteractor from './interactor';
 import mountComponent from './mountComponent';
-import mountComponentNotEditable from './mountComponent2';
 
-import { setupApplication } from '../../../tests/helpers';
+import {
+  setupApplication,
+  mount,
+} from '../../../tests/helpers';
+
+import connectStripes from '../../../tests/connectStripes';
+import ControlledVocab from '../ControlledVocab';
 
 describe('ControlledVocab', () => {
   const cv = new ControlledVocabInteractor();
@@ -124,9 +130,34 @@ describe('ControlledVocab', () => {
   });
 
   describe('User can edit EditableListForm', () => {
+    const ConnectedComponent = connectStripes(ControlledVocab);
     setupApplication();
 
-    beforeEach(mountComponentNotEditable);
+    beforeEach(() => {
+      mount(
+        <ConnectedComponent
+          baseUrl="location-units/institutions"
+          records="locinsts"
+          label="Institutions"
+          labelSingular="Institution"
+          objectLabel="Location"
+          visibleFields={['name', 'code', 'source']}
+          columnMapping={{
+            name: 'name',
+            code: 'code',
+            source: 'source',
+          }}
+          readOnlyFields={['source']}
+          itemTemplate={{ source: 'local' }}
+          hiddenFields={['description', 'numberOfObjects']}
+          nameKey="name"
+          // columnWidths={{ 'name': 300, 'code': 50 }}
+          id="institutions"
+          sortby="name"
+          editable={false}
+        />
+      );
+    });
 
     it('should render New button', () => {
       expect(cv.hasNewButton).to.be.false;
